extract retrieved_source_marks
